refactor: optimize json file
