+++ src/Files/JsonLanguage.cc
@@ -1,59 +1,15 @@
 #include "Global.h"
+#include <GMLIB/Files/JsonFile.h>
 #include <GMLIB/Files/JsonLanguage.h>
 
 namespace GMLIB::Files::JsonLanguage {
 
 nlohmann::json initLanguage(std::string path, nlohmann::json& defaultFile) {
-    writeOrUpdateFile(path, defaultFile);
-    return readFromFile(path);
-}
-
-void writeOrUpdateFile(std::string path, nlohmann::json& defaultFile) {
-    auto dirPath = std::filesystem::path(path).parent_path();
-    if (!std::filesystem::exists(dirPath)) {
-        std::filesystem::create_directories(dirPath);
-    }
-    if (!std::filesystem::exists(path)) {
-        writeFile(path, defaultFile);
-    } else {
-        updateFile(path, defaultFile);
-    }
+    return GMLIB::Files::JsonFile::initJson(path, defaultFile);
 }
 
 nlohmann::json initLanguage(std::string path, std::string& defaultFile) {
-    auto file = nlohmann::json::parse(defaultFile, nullptr, true, true);
-    return initLanguage(path, file);
-}
-
-void updateFile(std::string& path, nlohmann::json& newFile) {
-    auto oldFile = readFromFile(path);
-    newFile.merge_patch(oldFile);
-    writeFile(path, newFile);
-}
-
-void updateFile(std::string& path, std::string& newFile) {
-    auto file = nlohmann::json::parse(newFile, nullptr, true, true);
-    updateFile(path, file);
-}
-
-nlohmann::json readFromFile(std::string& path) {
-    std::ifstream inputFile(path);
-    std::string   fileContent((std::istreambuf_iterator<char>(inputFile)), std::istreambuf_iterator<char>());
-    return nlohmann::json::parse(fileContent, nullptr, true, true);
-}
-
-bool writeFile(std::string& path, nlohmann::json& json) {
-    try {
-        std::ofstream newFile(path);
-        if (!newFile.is_open()) {
-            return false;
-        }
-        newFile << json.dump(4);
-        newFile.close();
-        return true;
-    } catch (...) {
-        return false;
-    }
+    return GMLIB::Files::JsonFile::initJson(path, defaultFile);
 }
 
 std::string getValue(nlohmann::json& json, std::string key) {
@@ -74,19 +30,19 @@ std::string getValue(nlohmann::json& json, std::string key, std::string defaultV
 
 bool setValue(nlohmann::json& json, std::string key, std::string value, std::string path) {
     json[key] = value;
-    return writeFile(path, json);
+    return JsonFile::writeFile(path, json);
 }
 
 bool deleteKey(nlohmann::json& json, std::string key, std::string path) {
     json.erase(key);
-    return writeFile(path, json);
+    return JsonFile::writeFile(path, json);
 }
 
 bool deleteKeys(nlohmann::json& json, std::vector<std::string> keys, std::string path) {
     for (auto& key : keys) {
         json.erase(key);
     }
-    return writeFile(path, json);
+    return JsonFile::writeFile(path, json);
 }
 
 std::string translate(nlohmann::json& json, std::string key, std::vector<std::string> data, std::string translateKey) {
